sinks:vector:refactor
