+++ src/sinks/vector/push/compression.rs
